refactor(StringCalculator): Splitter 클래스 분리
